requirements.txt
natasha==1.6.0
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-cov==4.1.0
loguru==0.7.2 